--- index.css
@@ -1,78 +1,0 @@
-html, body {
-  height: 100%;
-  margin: 0;
-}
-
-main {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.row {
-  display: flex;
-}
-
-.top-row, .bottom-row {
-  margin: 10px 20px;
-  justify-content: space-between;
-  align-items: center;
-}
-
-#top-left {
-  display: flex;
-  justify-content: space-around;
-}
-
-#top-left div {
-  margin: 0px 10px;
-}
-
-#top-right {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-#top-right > div {
-  margin: 0px 0px;
-}
-
-.middle-row {
-  justify-content: center;
-  align-items: center;
-}
-
-#center > div {
-  text-align: center;
-}
-
-#bottom-left {
-  display: flex;
-}
-
-#bottom-left > div {
-  margin: 0px 10px;
-}
-
-#bottom-left, #bottom-right {
-  flex-grow: 1;
-  flex-basis: 25%;
-}
-
-#bottom-right {
-  display: flex;
-  justify-content: flex-end;
-}
-
-#bottom-center {
-  text-align: center;
-  flex-grow: 2;
-  flex-basis: 50%;
-}
-
-.icon-top-left {
-  width: 18px;
-  height: 18px;
-}
