+++ src/index.html
@@ -5,10 +5,7 @@
   <title></title>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="styles/index.css" rel="stylesheet">
-  <link href="styles/weather-icons.min.css" rel="stylesheet">
-  <script src="jquery-3.2.1.min.js"></script>
-  <script src="index.js"></script>
+  <script src="bundle.js"></script>
 </head>
 
 <body>
@@ -17,8 +14,8 @@
     <div class="row top-row">
       <div id="top-left">
         <div>Links</div>
-        <div><a href="#" id="chrome-tab-link" title="Chrome Tab"><img class="icon-top-left" src="./icons/chrome-grey.svg" alt="Chrome Tab"></a></div>
-        <div><a href="#" id="chrome-apps-link" title="Apps"><img class="icon-top-left" src="./icons/nine-squares-grey.svg" alt="Chrome Apps"></a></div>
+        <div><a href="#" id="chrome-tab-link" title="Chrome Tab"><img class="icon-top-left" src="./assets/images/chrome-grey.svg" alt="Chrome Tab"></a></div>
+        <div><a href="#" id="chrome-apps-link" title="Apps"><img class="icon-top-left" src="./assets/images/nine-squares-grey.svg" alt="Chrome Apps"></a></div>
         <div>Search</div>
       </div>
       <div id="top-right">
