+++ src/manifest.json
@@ -4,14 +4,14 @@
   "description": "Momentum clone",
   "version": "1.0.0",
   "icons": {
-    "16": "./icons/sea-turtle16.png",
-    "128": "./icons/sea-turtle128.png"
+    "16": "./assets/images/sea-turtle16.png",
+    "128": "./assets/images/sea-turtle128.png"
   },
   "chrome_url_overrides": {
     "newtab": "index.html"
   },
   "browser_action": {
-    "default_icon": "./icons/sea-turtle16.png"
+    "default_icon": "./assets/images/sea-turtle16.png"
   },
   "content_security_policy": "script-src 'self' https://ipinfo.io/ https://hickory-office.glitch.me/weather ; object-src 'self'",
   "permissions": [
